Refactoring JC-679
