+++ libs/core/dtos/authorization-tokens.dto.ts
@@ -1,0 +1,9 @@
+/** Tokens for user authorization. */
+export type AuthorizationTokensDto = {
+
+	/** Token for refresh access token. */
+	readonly refresh: string;
+
+	/** Access token. */
+	readonly access: string;
+};
