+++ libs/core/models/server-error-attribute.ts
@@ -1,4 +1,4 @@
-/** */
+/** Error's attribute. The form field whose value caused the error. */
 export enum ServerErrorAttribute {
 	NonField = 'nonFieldErrors',
 	Email = 'email',
